% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manage_code.R
\name{CreateGlobalTemplate}
\alias{CreateGlobalTemplate}
\title{Create a default global template for a home directory
This should be run once when GrandeJatte is installed}
\usage{
CreateGlobalTemplate(home_directory)
}
\arguments{
\item{home_directory}{A file path}
}
\description{
Create a default global template for a home directory
This should be run once when GrandeJatte is installed
}
\examples{
CreateGlobalTemplate("/Users/asmith01/projects")
CreateGlobalTemplate("~/projects")
}
