% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manage_code.R
\name{CreateTemplate}
\alias{CreateTemplate}
\title{Create a default template}
\usage{
CreateTemplate(home_directory)
}
\arguments{
\item{home_directory}{A file path}
}
\description{
Create a default template
}
\examples{
CreateTemplate("/Users/asmith01/projects")
CreateTemplate("~/projects")
}
