% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manage_code.R
\name{OpenLastSeurat}
\alias{OpenLastSeurat}
\title{Opens the most recently saved Seurat using the SaveSeurat function}
\usage{
OpenLastSeurat()
}
\value{
A Seurat object
}
\description{
Opens the most recently saved Seurat using the SaveSeurat function
}
\examples{
OpenLastSeurat()
}
