% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manage_code.R
\name{SaveSeurat}
\alias{SaveSeurat}
\title{Saves a Seurat object to the predefined directory}
\usage{
SaveSeurat(seurat_obj, seurat_filename, most_recent = TRUE)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{seurat_filename}{A filename ending in .rds}

\item{most_recet}{A boolean}
}
\description{
Saves a Seurat object to the predefined directory
}
\examples{
SaveSeurat(seurat, "05_smith_diff_exp.rds", most_recent = TRUE)
SaveSeurat(DATA, "DATA_unfiltered.rds", most_recent = FALSE)
}
