% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manage_code.R
\name{SaveViz}
\alias{SaveViz}
\title{Saves a visualization}
\usage{
SaveViz(category, filename, plot = last_plot(), width = 7, height = 7)
}
\arguments{
\item{category}{A string}

\item{filename}{A file path with an extension}

\item{plot}{The plot to save, default is last_plot()}

\item{width}{A number}

\item{height}{A number}
}
\description{
Saves a visualization
}
\examples{
SaveViz("qc", "percent_mito.pdf", width = 10, height = 10)
SaveViz("other", "bar_plot.png", plot = p1)
}
