% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manage_code.R
\name{SaveVizCluster}
\alias{SaveVizCluster}
\title{Saves a clustering visualization}
\usage{
SaveVizCluster(filename, plot = last_plot(), width = 7, height = 7)
}
\arguments{
\item{filename}{A file path with an extension}

\item{plot}{The plot to save, default is last_plot()}

\item{width}{A number}

\item{height}{A number}
}
\description{
Saves a clustering visualization
}
\examples{
SaveVizCluster("umap_.pdf", width = 10, height = 10)
SaveVizCluster("tsne_plot.png", plot = p1)
}
