% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manage_code.R
\name{SaveVizDE}
\alias{SaveVizDE}
\title{Saves a differential expression visualization}
\usage{
SaveVizDE(filename, plot = last_plot(), width = 7, height = 7)
}
\arguments{
\item{filename}{A file path with an extension}

\item{plot}{The plot to save, default is last_plot()}

\item{width}{A number}

\item{height}{A number}
}
\description{
Saves a differential expression visualization
}
\examples{
SaveVizDE("top_genes.pdf", width = 10, height = 10)
SaveVizDE("downregulated_genes.png", plot = p1)
}
