% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manage_code.R
\name{SaveVizQC}
\alias{SaveVizQC}
\title{Saves a quality control visualization}
\usage{
SaveVizQC(filename, plot = last_plot(), width = 7, height = 7)
}
\arguments{
\item{filename}{A file path with an extension}

\item{plot}{The plot to save, default is last_plot()}

\item{width}{A number}

\item{height}{A number}
}
\description{
Saves a quality control visualization
}
\examples{
SaveVizQC("percent_mito.pdf", width = 10, height = 10)
SaveVizQC("bar_plot.png", plot = p1)
}
