% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_manage_code.R
\name{SetProject}
\alias{SetProject}
\title{Sets the current project}
\usage{
SetProject(project_title, home_directory)
}
\arguments{
\item{project_title}{The project title - this should be found in the global template file}
}
\description{
Sets the current project
}
\examples{
SetProject('granger')
}
